Move edit for graphs to JournalEntryDisplay component instead of DoughnutGraph component
